Working directory
This is a fix on the working directory issue, in which the old code
attempted to read the settings inside the working directory (could be
/Windows/System32, say) instead of the execution directory (the folder
where the exe file is located).

Some users might not feel this is an issue, but if another program
called TU, it would not work properly.